+++ src/main/java/com/raoulvdberge/refinedstorage/apiimpl/network/node/externalstorage/StorageItemCyclops.java
@@ -15,7 +15,7 @@ import org.cyclops.cyclopscore.tileentity.InventoryTileEntityBase;
 
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
-import java.util.*;
+import java.util.Arrays;
 import java.util.function.Supplier;
 
 public class StorageItemCyclops extends StorageItemExternal {
@@ -27,7 +27,11 @@ public class StorageItemCyclops extends StorageItemExternal {
     public StorageItemCyclops(NetworkNodeExternalStorage externalStorage) {
         this.externalStorage = externalStorage;
         this.opposite = externalStorage.getHolder().getDirection().getOpposite();
-        this.cyclopsInv = () -> (InventoryTileEntityBase) externalStorage.getFacingTile();
+        this.cyclopsInv = () -> {
+            TileEntity f = externalStorage.getFacingTile();
+
+            return f instanceof InventoryTileEntityBase ? (InventoryTileEntityBase) f : null;
+        };
     }
 
     @Override
@@ -87,9 +91,9 @@ public class StorageItemCyclops extends StorageItemExternal {
         if (inv != null) {
             if (inv.getInventory() instanceof IndexedSlotlessItemHandlerWrapper.IInventoryIndexReference) {
                 return ((IndexedSlotlessItemHandlerWrapper.IInventoryIndexReference) inv.getInventory())
-                        .getIndex().values().stream().flatMap(m -> m.valueCollection().stream()).map(ItemStack::copy).collect(RSUtils.toNonNullList());
+                    .getIndex().values().stream().flatMap(m -> m.valueCollection().stream()).map(ItemStack::copy).collect(RSUtils.toNonNullList());
             } else {
-                return Arrays.stream(((SimpleInventory)inv.getInventory()).getItemStacks()).map(ItemStack::copy).collect(RSUtils.toNonNullList());
+                return Arrays.stream(((SimpleInventory) inv.getInventory()).getItemStacks()).map(ItemStack::copy).collect(RSUtils.toNonNullList());
             }
         } else {
             return RSUtils.emptyNonNullList();
@@ -98,7 +102,7 @@ public class StorageItemCyclops extends StorageItemExternal {
 
     public static boolean isValid(TileEntity facingTE, EnumFacing facing) {
         return facingTE instanceof InventoryTileEntityBase
-                && (SlotlessItemHandlerHelper.isSlotless(facingTE, facing)
-                    || ((InventoryTileEntityBase) facingTE).getInventory() instanceof SimpleInventory);
+            && (SlotlessItemHandlerHelper.isSlotless(facingTE, facing)
+            || ((InventoryTileEntityBase) facingTE).getInventory() instanceof SimpleInventory);
     }
 }